+++ Kernel/Devices/Network/Intel/E1000ENetworkAdapter.h
@@ -9,9 +9,9 @@
 #include <AK/OwnPtr.h>
 #include <Kernel/Bus/PCI/Access.h>
 #include <Kernel/Bus/PCI/Device.h>
+#include <Kernel/Devices/Network/Intel/E1000NetworkAdapter.h>
 #include <Kernel/Interrupts/IRQHandler.h>
 #include <Kernel/Library/IOWindow.h>
-#include <Kernel/Net/Intel/E1000NetworkAdapter.h>
 #include <Kernel/Net/NetworkAdapter.h>
 #include <Kernel/Security/Random.h>
