+++ Kernel/Devices/Network/Intel/E1000NetworkAdapter.h
