+++ Kernel/Devices/Network/Realtek/RTL8168NetworkAdapter.cpp
@@ -9,8 +9,8 @@
 #include <Kernel/Bus/PCI/API.h>
 #include <Kernel/Bus/PCI/IDs.h>
 #include <Kernel/Debug.h>
+#include <Kernel/Devices/Network/Realtek/RTL8168NetworkAdapter.h>
 #include <Kernel/Net/NetworkingManagement.h>
-#include <Kernel/Net/Realtek/RTL8168NetworkAdapter.h>
 #include <Kernel/Sections.h>
 
 namespace Kernel {
