+++ Meta/gn/secondary/Kernel/BUILD.gn
@@ -548,6 +548,10 @@ executable("Kernel_bin") {
     "Devices/HID/PS2/MouseDevice.cpp",
     "Devices/KCOVDevice.cpp",
     "Devices/KCOVInstance.cpp",
+    "Devices/Network/Intel/E1000ENetworkAdapter.cpp",
+    "Devices/Network/Intel/E1000NetworkAdapter.cpp",
+    "Devices/Network/Realtek/RTL8168NetworkAdapter.cpp",
+    "Devices/Network/VirtIO/VirtIONetworkAdapter.cpp",
     "Devices/Serial/VirtIO/Console.cpp",
     "Devices/Serial/VirtIO/ConsolePort.cpp",
     "Devices/SerialDevice.cpp",
@@ -698,19 +702,15 @@ executable("Kernel_bin") {
     "Memory/VMObject.cpp",
     "Memory/VirtualRange.cpp",
     "Net/IPv4Socket.cpp",
-    "Net/Intel/E1000ENetworkAdapter.cpp",
-    "Net/Intel/E1000NetworkAdapter.cpp",
     "Net/LocalSocket.cpp",
     "Net/LoopbackAdapter.cpp",
     "Net/NetworkAdapter.cpp",
     "Net/NetworkTask.cpp",
     "Net/NetworkingManagement.cpp",
-    "Net/Realtek/RTL8168NetworkAdapter.cpp",
     "Net/Routing.cpp",
     "Net/Socket.cpp",
     "Net/TCPSocket.cpp",
     "Net/UDPSocket.cpp",
-    "Net/VirtIO/VirtIONetworkAdapter.cpp",
     "SanCov.cpp",
     "Security/AddressSanitizer.cpp",
     "Security/Credentials.cpp",
